requirements.txt
algorithmia==1.0.8
amqp==1.4.9
aniso8601==1.2.1
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.3.1
asn1crypto==0.22.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-bootstrap-form==3.2.1
django-bootstrap-ui==0.4.0
django-bootstrap3==8.2.3
django-celery==3.2.1
django-crispy-forms==1.6.1
django-rest-swagger==2.1.2
django-tag-parser==2.1
django-tastypie==0.14.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
dominate==2.1.17
enum34==1.1.6
Flask==0.10.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.12
funcsigs==1.0.2
futures==3.1.1
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.8
kombu==3.0.37
MarkupSafe==0.23
mysqlclient==1.3.10
nltk==3.2.4
numpy==1.12.1
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
pandas==0.20.2
Pillow==4.1.1
pycparser==2.17
pyparsing==2.2.0
pysolr==3.4.0
python-craigslist==1.0.3
python-dateutil==2.6.0
python-decouple==3.1
python-mimeparse==1.6.0
pytz==2017.2
requests==2.9.1
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.9
tzlocal==1.4
uritemplate==3.0.0
urllib3==1.21.1
watson-developer-cloud==0.6.0
Werkzeug==0.11.4
whitenoise==3.3.0
WTForms==2.1
